pyproject.toml
[project]
name = "scansnap-web"
version = "0.1.0"
description = "ScanSnap web application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "httpx>=0.28.1",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "pytest>=8.3.5",
    "ruff>=0.11.7",
]
